+++ clicache/src/UserFunctionExecutionException.hpp
@@ -99,9 +99,9 @@ namespace Apache
         /// <exception cref="IllegalStateException">
         /// If this api is called from User code.
         /// </exception>
-        virtual property System::UInt32 ObjectSize
+        virtual property System::UInt64 ObjectSize
         {
-          virtual System::UInt32 get();
+          virtual System::UInt64 get();
         }
 
         // End: IGeodeSerializable members
