+++ tests/cli/NewTestObject/PositionPdx.cs
@@ -68,7 +68,7 @@ namespace Apache.Geode.Client.Tests
       m_pid = 0;
     }
 
-    private UInt32 GetObjectSize(IGeodeSerializable obj)
+    private UInt64 GetObjectSize(IGeodeSerializable obj)
     {
       return (obj == null ? 0 : obj.ObjectSize);
     }
